mido: ConsumerIr: Add Dynamic Lifecycle HAL support
IR HAL is rarely in use practically,but it always runs in background &
consumes power/resources as well.

Android Q adds a new callback and helper functions to make it easier to
make use of dynamic
HALs , which were added in P. Dynamic HALs allow you to start your HAL
when needed, and
either exit or reduce resource consumption when it’s not being used.

https://source.android.com/devices/architecture/hal/dynamic-lifecycle

Signed-off-by: ShihabZzz <Shihab.riadn@gmail.com>
Change-Id: Id116293e4edd3dc0e64ebcdec60357da92b0b3e6
Signed-off-by: smokey18 <kmboplays@gmail.com>